@@ -1,9 +1,9 @@
 package com.example.abhishek.newsapp.network;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.example.abhishek.newsapp.models.Article;
 import com.example.abhishek.newsapp.models.ArticleResponseWrapper;
