@@ -1,9 +1,9 @@
 package com.example.abhishek.newsapp.ui.news;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.annotation.NonNull;
 
 import com.example.abhishek.newsapp.data.NewsRepository;
 import com.example.abhishek.newsapp.models.Article;
