@@ -1,7 +1,7 @@
 package com.example.abhishek.newsapp.utils;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.DisplayMetrics;
 
 import com.bumptech.glide.annotation.GlideModule;
