@@ -1,18 +1,19 @@
 package com.example.abhishek.newsapp.ui.sources;
 
-import android.arch.lifecycle.Observer;
-import android.arch.lifecycle.ViewModelProviders;
 import android.content.Intent;
-import android.databinding.DataBindingUtil;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.DividerItemDecoration;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.databinding.DataBindingUtil;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.recyclerview.widget.DividerItemDecoration;
 
 import com.example.abhishek.newsapp.R;
 import com.example.abhishek.newsapp.adapters.SourceAdapter;
@@ -62,7 +63,7 @@ public class SourceFragment extends Fragment implements SourceAdapter.SourceAdap
         Specification specification = new Specification();
         specification.setLanguage(Locale.getDefault().getLanguage());
         specification.setCountry(null);
-        viewModel.getSource(specification).observe(this, new Observer<List<Source>>() {
+        viewModel.getSource(specification).observe(getViewLifecycleOwner(), new Observer<List<Source>>() {
             @Override
             public void onChanged(@Nullable List<Source> sources) {
                 if (sources != null) {
